Finish refactor the code
