+++ app/.eslintrc.json
@@ -5,7 +5,8 @@
     "node/exports-style": ["error", "module.exports"],
     "arrow-parens": ["error", "as-needed"],
     "no-underscore-dangle": ["error", { "allow": ["_id", "_index"] }],
-    "function-paren-newline": ["error", { "minItems": 3 }]
+    "function-paren-newline": ["error", "multiline"],
+    "max-len": ["error", { "code": 160 }]
   },
   "env": { "es6": true },
   "parserOptions": {
