+++ app/bin/www
@@ -4,6 +4,8 @@
  * Module dependencies.
  */
 
+/* eslint-disable */
+
 var app = require('../app');
 var debug = require('debug')('app:server');
 var http = require('http');
