+++ app/index.js
@@ -3,6 +3,6 @@ const express = require('express');
 const router = express.Router();
 
 /* GET home page. */
-router.get('/', (req, res) => res.send('Hello World!'));
+router.get('/', (req, res) => res.send('Server is up!'));
 
 module.exports = router;
