+++ app/logger.js
@@ -19,11 +19,11 @@ const logger = winston.createLogger({
 // `${info.level}: ${info.message} JSON.stringify({ ...rest }) `
 //
 if (process.env.NODE_ENV !== 'production') {
-  logger.add(
-    new winston.transports.Console({
-      format: winston.format.simple(),
-    }),
-  );
+  const format = {
+    format: winston.format.simple(),
+  };
+
+  logger.add(new winston.transports.Console(format));
 }
 
 module.exports = logger;
